Move some types to their own headers.
