+++ src/include/smol/smol_renderer_types.h
@@ -10,6 +10,9 @@
 #include <smol/smol_vector4.h>
 #include <smol/smol_mat4.h>
 #include <smol/smol_camera.h>
+#include <smol/smol_material.h>
+#include <smol/smol_shader.h>
+#include <smol/smol_texture.h>
 
 #define SMOL_GL_DEFINE_EXTERN
 #include <smol/smol_gl.h> //TODO(marcio): Make this API independent. Remove all GL specifics from this header
@@ -69,128 +72,6 @@ namespace smol
     TRIANGLE_STRIP,
     LINE,
     POINT
-  };
-
-
-  struct SMOL_ENGINE_API Texture
-  {
-    enum Wrap
-    {
-      REPEAT            = 0,
-      REPEAT_MIRRORED   = 1,
-      CLAMP_TO_EDGE     = 2,
-      MAX_WRAP_OPTIONS
-    };
-
-    enum Filter
-    {
-      LINEAR                  = 0,
-      NEAREST                 = 1,
-      MAX_FILTER_OPTIONS
-    };
-
-    enum Mipmap
-    {
-      LINEAR_MIPMAP_LINEAR    = 0,
-      LINEAR_MIPMAP_NEAREST   = 1,
-      NEAREST_MIPMAP_LINEAR   = 2,
-      NEAREST_MIPMAP_NEAREST  = 3,
-      NO_MIPMAP               = 4,
-      MAX_MIPMAP_OPTIONS
-    };
-
-    int width;
-    int height;
-    GLuint textureObject;
-  };
-
-#define SMOL_MAX_SHADER_PARAMETER_NAME_LEN 64
-  struct ShaderParameter
-  {
-    public:
-    enum Type
-    {
-      SAMPLER_2D,
-      VECTOR2,
-      VECTOR3,
-      VECTOR4,
-      FLOAT,
-      INT,
-      UNSIGNED_INT,
-      INVALID
-    };
-
-    Type type;
-    char name[SMOL_MAX_SHADER_PARAMETER_NAME_LEN];
-    GLuint location;
-  };
-
-  struct SMOL_ENGINE_API MaterialParameter : public ShaderParameter
-  {
-    union
-    {
-      float floatValue;
-      Vector2 vec2Value;
-      Vector3 vec3Value;
-      Vector4 vec4Value;
-      int32 intValue;
-      uint32 uintValue;
-    };
-  };
-
-#define SMOL_MAX_SHADER_PARAMETERS 16
-  struct SMOL_ENGINE_API ShaderProgram
-  {
-    bool valid;
-    GLuint programId;
-    ShaderParameter parameter[SMOL_MAX_SHADER_PARAMETERS];
-    int parameterCount;
-  };
-
-#define SMOL_MATERIAL_MAX_TEXTURES 6
-#define SMOL_MAX_BUFFERS_PER_MESH 6
-  struct SMOL_ENGINE_API Material
-  {
-    enum DepthTest
-    {
-      DISABLE         = 0,
-      LESS            = 1,
-      LESS_EQUAL      = 2,
-      EQUAL           = 3,
-      GREATER         = 4,
-      GREATER_EQUAL   = 5,
-      DIFFERENT       = 6,
-      NEVER           = 7,
-      ALWAYS          = 8
-    };
-
-    enum CullFace
-    {
-      BACK            = 0,
-      FRONT           = 1,
-      FRONT_AND_BACK  = 2,
-      NONE            = 3
-    };
-
-    Handle<ShaderProgram> shader;
-    Handle<Texture> textureDiffuse[SMOL_MATERIAL_MAX_TEXTURES];
-    int diffuseTextureCount;
-    int renderQueue;
-    MaterialParameter parameter[SMOL_MAX_SHADER_PARAMETERS];
-    int parameterCount;
-    DepthTest depthTest;
-    CullFace cullFace;
-
-    Material& setSampler2D(const char* name, unsigned int value);
-    Material& setUint(const char* name, unsigned int value);
-    Material& setInt(const char* name, int value);
-    Material& setFloat(const char* name, float value);
-    Material& setVec2(const char* name, const Vector2& value);
-    Material& setVec3(const char* name, const Vector3& value);
-    Material& setVec4(const char* name, const Vector4& value);
-
-    private:
-    MaterialParameter* getParameter(const char* name, ShaderParameter::Type type);
   };
 
   struct SMOL_ENGINE_API Mesh
